[filter-layout] Decomposed map on map service
