+++ price-map/apps/frontend/src/app/modules/map/map.module.ts
@@ -2,12 +2,14 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MapComponent } from './components';
 import { MapRoutingModule } from './map-routing.module';
+import { MapService } from './services';
 
 @NgModule({
   declarations: [MapComponent],
   imports: [
     CommonModule,
     MapRoutingModule
-  ]
+  ],
+  providers: [MapService]
 })
 export class MapModule { }
