+++ price-map/apps/frontend/src/app/modules/map/services/index.ts
@@ -1,0 +1,1 @@
+export { MapService } from './map.service';
